fix(module5-task1): Вытащил useMap в отлдельный файл
